Career Path Recommendations(Career Counseling Chatbot)/requirements.txt
streamlit
google-generativeai 
python-dotenv
